Abstract.AdhesiveHLR

Documentation

class Eq m => Morphism m where

Minimal complete definition

compose, domain, codomain, id, isMonomorphism, isEpimorphism, isIsomorphism

Associated Types

type Obj m :: *

Methods

compose :: m -> m -> m

Apply the first argument first (compose f g = g . f)

domain :: m -> Obj m

codomain :: m -> Obj m

id :: Obj m -> m

isMonomorphism :: m -> Bool

isEpimorphism :: m -> Bool

isIsomorphism :: m -> Bool

class Morphism m => EpiPairs m where

Minimal complete definition

createJointlyEpimorphicPairs, createAllSubobjects, createJointlyEpimorphicPairsFromNAC, calculateCommutativeSquaresAlongMonomorphism

Methods

createJointlyEpimorphicPairs :: Bool -> Obj m -> Obj m -> [(m, m)]

Create all jointly epimorphic pairs of morphisms from the given objects.

If the first argument is true, only pairs of monomorphisms are created. Otherwise, pairs of arbitrary morphisms are created.

createAllSubobjects :: Bool -> Obj m -> [m]

createJointlyEpimorphicPairsFromNAC :: DPOConfig -> Obj m -> m -> [(m, m)]

Create a special case of jointly epimorphic pairs, where the second morphism is a Nac The first flag indicates Nac satisfability with a monomorphic morphism The second flag indicates that the other morphism is monomorphic

FIXME: nacs don't belong in this module

calculateCommutativeSquares :: Bool -> m -> m -> [(m, m)]

calculateCommutativeSquaresAlongMonomorphism :: (m, Bool) -> (m, Bool) -> [(m, m)]

class Morphism m => AdhesiveHLR m where

Type class for morphisms whose category Adhesive and suitable for High-Level Replacement Systems.

Mainly provides categorical operations that AdhesiveHLR categories are guaranteed to have.

Minimal complete definition

calculatePushout, hasPushoutComplement, calculatePushoutComplement, monomorphicPullback

Methods

calculatePushout :: m -> m -> (m, m)

Calculate the pushout between the two given morphisms.

Given the morphisms f : A -> B and g : A -> C, respectively, returns the pair of morphisms f' : C -> D and g': B -> D such that the following square is a pushout.

      g
   A──────▶C
   │       │
 f │       │ f'
   ▼       ▼
   B──────▶D
      g'

hasPushoutComplement :: (MorphismType, m) -> (MorphismType, m) -> Bool

Checks if the given sequential morphisms have a pushout complement, assuming they satsify the given restriction.

Given the morphisms g : B -> C and f : A -> B, respectively, tests if there exists a pair of morphisms f' : A -> X and g' : X -> B such that the following square is a pushout. Since the category is Adhesive, such a pair is unique.

       f
    A──────▶B
    │       │
 g' │       │ g
    ▼       ▼
    X──────▶C
       f'

If the types of the morphisms are known, they should be given. The implementation of this operation may then use them for more efficient calculation.

calculatePushoutComplement :: m -> m -> (m, m)

Calculate the pushout complement for two sequential morphisms, assumes it exists.

In order to test if the pushout complement exists, use hasPushoutComplement.

Given the morphisms g : B -> C and f : A -> B, respectively, returns the pair of morphisms f' : A -> X and g' : X -> B such that the following square is a pushout. Since the category is Adhesive, such a pair is unique.

       f
    A──────▶B
    │       │
 g' │       │ g
    ▼       ▼
    X──────▶C
       f'

monomorphicPullback :: m -> m -> (m, m)

Calculate the pullback between the two given morphisms

Given two monomorphic morphisms f : A -> C and g : B -> C, respectively, returns the pair of monomorphic morphisms f' : X -> B and g': X -> A such that the following square is a pullback.

       g'
    X──────▶A
    │       │
 f' │       │ f
    ▼       ▼
    B──────▶C
       g

data MatchRestriction

Flag indicating what restrictions are required or assumed of matches.

Constructors

MonoMatches

AnyMatches

matchRestrictionToMorphismType :: MatchRestriction -> MorphismType

Converts a match restriction to the corresponding MorphismType

data NacSatisfaction

Flag indicating the semantics of NAC satisfaction.

Constructors

MonomorphicNAC

PartiallyMonomorphicNAC

data DPOConfig

Constructors

DPOConfig

Fields

matchRestriction :: MatchRestriction

nacSatisfaction :: NacSatisfaction