Abstract.Constraint

Documentation

data AtomicConstraint m

Constructors

AtomicConstraint

Fields

name :: String

morphism :: m

positive :: Bool

buildNamedAtomicConstraint :: String -> m -> Bool -> AtomicConstraint m

satisfiesAtomicConstraint :: FindMorphism m => Obj m -> AtomicConstraint m -> Bool

Given an object G and a AtomicConstraint a : P -> C, check whether G satisfies the AtomicConstraint a

satisfiesAllAtomicConstraints :: FindMorphism m => Obj m -> [AtomicConstraint m] -> Bool

Given an object G and a list of AtomicConstraints a : P -> C, check whether G satisfies the all them

data Constraint m

Constructors

Atomic

Fields

atomic :: AtomicConstraint m

And

Fields

lc :: Constraint m

rc :: Constraint m

Or

Fields

lc :: Constraint m

rc :: Constraint m

Not

Fields

nc :: Constraint m

satisfiesConstraint :: FindMorphism m => Obj m -> Constraint m -> Bool

Given an object G and a Constraint c (a Boolean formula over atomic constraints), check whether G satisfies c

satisfiesAllConstraints :: FindMorphism m => Obj m -> [Constraint m] -> Bool

Given an object G and a list of Constraints (Boolean formulas over atomic constraints), check whether G satisfies the all them