Abstract.DPO.Process

Documentation

data Process m

Constructors

Process

Fields

productions :: [Production m]

coreObject :: Obj m

class DPO m => GenerateProcess m where

Minimal complete definition

restrictMorphisms, restrictMorphism, typing, productionTyping

Methods

restrictMorphisms :: (m, m) -> (m, m)

Given a pair of morhisms with common codomain, it returns a new pair with morphism also with a a new codomain that does not contain the elements that were orphans in both original morphisms

restrictMorphism :: m -> m

Given a morhism, it returns a morphism with a new codomain that is equal to the image of the original morphism

typing :: (Derivation m, (m, m, m)) -> Production m

Given a Derivation d and a tuple (p,q,r) of Morphisms p : G -> C, q : D -> C and r : H -> C, it returns a new Production corresponding to the production in d but typed over C

productionTyping :: (Production m, (m, m, m)) -> Production m

Given a Production p and a tuple (r,s,t) of Morphisms r : G -> C, s : D -> C and t : H -> C, it returns a new Production corresponding to the production in p but typed over C

calculateProcess :: [Derivation m] -> Process m

Given a list of Derivation containing a sequential derivation, returns its corresponding Process

calculateRulesColimit :: RuleSequence m -> [NamedRuleWithMatches m]

generateGraphProcess :: RuleSequence m -> [(String, Production m)]

type NamedRuleWithMatches m = (String, Production m, (m, m, m))

data Interaction

Constructors

Interaction

Fields

firstRule :: String

secondRule :: String

interactionType :: InteractionType

nacInvolved :: Maybe Int

data InteractionType

Constructors

DeleteUse

ProduceForbid

ProduceUse

DeleteForbid

getRule :: NamedRuleWithMatches m -> Production m

getName :: NamedRuleWithMatches m -> String

getMatch :: NamedRuleWithMatches m -> m

getComatch :: NamedRuleWithMatches m -> m

filterInducedByNacs :: [Interaction] -> Set Interaction

eliminateSelfConflictsAndDependencies :: [Interaction] -> [Interaction]