Abstract.Morphism

Documentation

class Eq m => Morphism m where

Minimal complete definition

compose, domain, codomain, id, isMonomorphism, isEpimorphism, isIsomorphism

Associated Types

type Obj m :: *

Methods

compose :: m -> m -> m

Apply the first argument first (compose f g = g . f)

domain :: m -> Obj m

codomain :: m -> Obj m

id :: Obj m -> m

isMonomorphism :: m -> Bool

isEpimorphism :: m -> Bool

isIsomorphism :: m -> Bool

data MorphismType

Enum for the types of morphisms that can be used / found

Constructors

GenericMorphism

Monomorphism

Epimorphism

Isomorphism

class Morphism m => FindMorphism m where

Minimal complete definition

findMorphisms, partialInjectiveMatches

Methods

findMorphisms :: MorphismType -> Obj m -> Obj m -> [m]

Given a type t of MorphismType and two objects A and B, it finds all the matches m : A -> B in which m is of the type t

findMonomorphisms :: Obj m -> Obj m -> [m]

Given two objects A and B, finds all monomorphisms from A to B

findEpimorphisms :: Obj m -> Obj m -> [m]

Given two objects A and B, finds all epimorphisms from A to B

findIsomorphisms :: Obj m -> Obj m -> [m]

Given two objects A and B, finds all isomorphisms from A to B

findAllMorphisms :: Obj m -> Obj m -> [m]

Given two objects A and B, finds all morphisms from A to B

partialInjectiveMatches :: m -> m -> [m]

Given a NAC n : L -> N and a match m : L -> G, finds the morphisms from N to G that are injective out of the image of n