Abstract.Relation

Types

data Relation a

Datatype for endorelations on a

Construction

empty :: Ord a => [a] -> [a] -> Relation a

An empty relation, with domain and codomain specified.

Transformation

compose :: Ord a => Relation a -> Relation a -> Relation a

Compose r1 and r2.

id :: Ord a => [a] -> Relation a

The identity relation on dom.

inverseRelation :: Ord a => Relation a -> Relation a

The inverse relation.

updateRelation :: Ord a => a -> a -> Relation a -> Relation a

Add a mapping between x and y to the relation. If x already exists, y is joined to the corresponding elements.

removeFromDomain :: Ord a => a -> Relation a -> Relation a

Remove an element from the domain of the relation

removeFromCodomain :: Ord a => a -> Relation a -> Relation a

Remove an element from the codomain of the relation

insertOnCodomain :: Ord a => a -> Relation a -> Relation a

Insert an element on the codomain of the relation

Query

apply :: Ord a => Relation a -> a -> [a]

Return a list of all elements that x gets mapped into.

domain :: Relation a -> [a]

codomain :: Relation a -> [a]

image :: Eq a => Relation a -> [a]

Return a list of all elements in the image of the relation.

mapping :: Relation a -> Map a [a]

orphans :: Eq a => Relation a -> [a]

Return the elements in the domain which are not in the image of the relation (orphans)

Predicates

isFunctional :: Relation a -> Bool

Test if r is functional.

isInjective :: Ord a => Relation a -> Bool

Test if r is injective.

isPartialInjective :: Ord a => [a] -> Relation a -> Bool

Test if r is injective out of domain list

isSurjective :: Ord a => Relation a -> Bool

Test if r is surjective.

isTotal :: Ord a => Relation a -> Bool

Test if r is total.