Abstract.Valid

Documentation

class Valid a where

Type class for types that admit "malformed" values, and must thus be checked.

Although implementing the isValid method is sufficient, it will result in poor error messages. Therefore, implementing the validate method is recommended.

Minimal complete definition

isValid | validate

Methods

isValid :: a -> Bool

Checks if the given value is well-formed

validate :: a -> ValidationResult

Checks if the given value is well-formed, providing an explanation of any errors encountered.

data ValidationResult

Result of validating some value

ensure :: Bool -> String -> ValidationResult

Return IsValid if the given boolean is true, otherwise return IsInvalid with the given error message.

withContext :: String -> ValidationResult -> ValidationResult

Prepends the indication of context to any error messages produced by the validation result.

errorMessages :: ValidationResult -> Maybe String

Obtains the error messages produced by validation, or Nothing if no error was found

validateNamed :: Valid a => (name -> String) -> [(name, a)] -> ValidationResult

Validates a list of named values, modifying their names with the given function.

ensureValid :: ValidationResult -> IO ()

If the given validation detected some error, prints the error out and exits with failure.