Analysis.ConcurrentRules

Documentation

data CRDependencies

Constructors

AllOverlapings

OnlyDependency

allConcurrentRules :: (DPO m, EpiPairs m, Eq (Obj m), Valid m) => CRDependencies -> DPOConfig -> [AtomicConstraint m] -> [Production m] -> [Production m]

Generates the Concurrent Rules for a given list of GraphRules following the order of the elements in the list.

maxConcurrentRule :: (DPO m, EpiPairs m, Eq (Obj m), Valid m) => CRDependencies -> DPOConfig -> [AtomicConstraint m] -> [Production m] -> Maybe (Production m)

Generates the Concurrent Rule with the least disjoint EpiPair for a given list of GraphRules (following the order of the elements in the list).