Analysis.CriticalPairs

Documentation

data CriticalPairType

Data representing the type of a CriticalPair

Constructors

FreeOverlap

DeleteUse

ProduceForbid

ProduceDangling

data CriticalPair m

A Critical Pair is defined as two matches (m1,m2) from the left side of their rules to a same graph. It assumes that the derivation of the rule with match m1 causes a conflict with the rule with match m2

This diagram shows graphs and morphisms names used in the algorithms below

p1 = production (L1,K1,R1,[N1])

p2 = production (L2,K2,R2,[N2])

                   N1    N2
                   ^      ^
         r     l   │      │n
    R1◀─────K1────▶L1    L2◀────K2─────▶R2
    │       │       \   /       │       │
 m1'│      k│      m1\ /m2      │       │
    ▼       ▼         ▼         ▼       ▼
    P1◀─────D1───────▶G◀───────D2──────▶P2
        r'       l'

m2' :: from L2 to P1

h21 :: from L2 to D1

q21 (nacMatch) :: from N2 to P1

Constructors

CriticalPair

Fields

matches :: (m, m)

comatches :: Maybe (m, m)

nacMatch :: Maybe (m, Int)

cpType :: CriticalPairType

getCriticalPairMatches :: CriticalPair m -> (m, m)

Returns the matches (m1,m2)

getCriticalPairComatches :: CriticalPair m -> Maybe (m, m)

Returns the comatches (m1',m2')

getNacIndexOfCriticalPair :: CriticalPair m -> Maybe Int

Returns the nac index of a CriticalPair

getNacMatchOfCriticalPair :: CriticalPair m -> Maybe m

Returns the nac match of a CriticalPair

getCriticalPairType :: CriticalPair m -> CriticalPairType

Returns the type of a Critical Pair

Finding Critical Pairs

findCriticalPairs :: (EpiPairs m, DPO m) => MorphismsConfig -> Production m -> Production m -> [CriticalPair m]

Finds all Critical Pairs between two given Productions

findPotentialCriticalPairs :: (DPO m, EpiPairs m) => MorphismsConfig -> Production m -> Production m -> [(m, m)]

Returns a list of morphisms from left side of rules to all valid overlapping pairs

namedCriticalPairs :: (EpiPairs m, DPO m) => MorphismsConfig -> [NamedRule m] -> [NamedCriticalPairs m]

Returns the Critical Pairs with rule names

findAllDeleteUse :: (EpiPairs m, DPO m) => MorphismsConfig -> Production m -> Production m -> [CriticalPair m]

All DeleteUse caused by the derivation of p1 before p2. It occurs when p1 deletes something used by p2.

findAllProduceForbid :: (EpiPairs m, DPO m) => MorphismsConfig -> Production m -> Production m -> [CriticalPair m]

All ProduceForbid caused by the derivation of p1 before p2.

Rule p1 causes a produce-forbid conflict with p2 if some NAC in p2 fails to be satisfied after the aplication of p1.

findAllProduceDangling :: (EpiPairs m, DPO m) => MorphismsConfig -> Production m -> Production m -> [CriticalPair m]

All ProduceDangling caused by the derivation of p1 before p2. It occurs when p1 creates something that unable p2.

findAllDeleteUseAndProduceDangling :: (EpiPairs m, DPO m) => MorphismsConfig -> Production m -> Production m -> [CriticalPair m]

Tests DeleteUse and ProduceDangling for the same pairs, more efficient than deal separately.