Analysis.Interlevel.EvolutionarySpans

Description

Documentation

allEvolSpans :: MorphismsConfig -> [(String, SndOrderRule a b)] -> [(String, String, [EvoSpan a b])]

Given a list of second order rules, calculate all Evolutionary Spans This analysis is supposed to be symmetric, here is considering only this case

data EvoSpan a b

Represents the two evolutionary matches, and the kind of span

leftMatch :: EvoSpan a b -> RuleMorphism a b

rightMatch :: EvoSpan a b -> RuleMorphism a b

cpe :: EvoSpan a b -> CPE