Dot

Documentation

data NamingContext

Constructors

Ctx

Fields

getNodeTypeName :: NodeId -> String

getEdgeTypeName :: EdgeId -> String

printDigraph :: String -> [String] -> [Doc] -> [Doc] -> Doc

printStateSpace :: StateSpace (TypedGraphMorphism a b) -> Doc

Create a dotfile representation of the given state space, labeling states with their IDs

printTypedGraph :: NamingContext -> String -> TypedGraph a b -> Doc

Create a dotfile representation of the given typed graph, labeling nodes with their types

nodeId :: NodeId -> Doc

edgeId :: EdgeId -> Doc