Grammar.Core

Description

A grammar is defined as a start object, a set of transformation rules, and a set of constraints for the potencial generated objects. The morphism type is kept generic.

Documentation

grammar :: Obj m -> [Constraint m] -> [NamedProduction m] -> Grammar m

data Grammar m

data ObjectFlow m

Object that uses a Span of Morphisms to connect the right-hand-side of a Production with the left-hand-side of another one

Constructors

ObjectFlow

Fields

index :: String

A identifier for the Object Flow

producer :: String

The name of the production that will produce the input for the next

consumer :: String

The name of the production that uses the result of the other

spanMapping :: Span m

A span of Morphisms Ri IO - Lo where Ri is the right-hand-side of the producer production and Lo is the left-hand-side of the consumer production

type RuleSequence m = (String, [(String, Production m)], [ObjectFlow m])

type NamedProduction m = (String, Production m)

getProductionName :: NamedProduction m -> String

getProduction :: NamedProduction m -> Production m

start :: Grammar m -> Obj m

constraints :: Grammar m -> [Constraint m]

rules :: Grammar m -> [NamedProduction m]

findProduction :: String -> Grammar m -> Maybe (Production m)

reachableGraphs :: Grammar m -> [(String, Obj m)]

addReachableGraphs :: [(String, Obj m)] -> Grammar m -> Grammar m