Image.Dot

Documentation

data NamingContext

Constructors

Ctx

Fields

getNodeTypeName :: NodeId -> String

getEdgeTypeName :: EdgeId -> String

makeNamingContext :: [(String, String)] -> NamingContext

printDotGraph :: String -> String -> [String] -> [Doc] -> [Doc] -> [Doc] -> Doc

printDigraph :: String -> [String] -> [Doc] -> [Doc] -> [Doc] -> Doc

printSubgraph :: String -> [String] -> [Doc] -> [Doc] -> [Doc] -> Doc

printStateSpace :: StateSpace (TypedGraphMorphism a b) -> Doc

Create a dotfile representation of the given state space, labeling states with their IDs

printTypedGraph :: NamingContext -> String -> TypedGraph a b -> Doc

Create a dotfile representation of the given typed graph, labeling nodes with their types

printSubTypedGraph :: NamingContext -> String -> TypedGraph a b -> Doc

printTypedGraphMorphism :: NamingContext -> String -> TypedGraphMorphism a b -> Doc

Create a dotfile representation of the given typed graph morphism

printGraphRule :: NamingContext -> String -> GraphRule a b -> Doc

Create a dotfile representation of the given graph rule

printSubGraphRule :: NamingContext -> String -> GraphRule a b -> Doc

printSndOrderRule :: NamingContext -> String -> SndOrderRule a b -> Doc

Create a dotfile representation of the given snd order rule

printNode :: Doc -> [Doc] -> Doc

printEdge :: Doc -> Doc -> [Doc] -> Doc

nodeSubId :: String -> NodeId -> Doc

nodeId :: NodeId -> Doc

edgeSubId :: String -> EdgeId -> Doc

edgeId :: EdgeId -> Doc