Logic.Ctl

Description

CTL Expressions

data Expr

CTL expressions

Constructors

Literal Bool

Atom String

Not Expr

And Expr Expr

Or Expr Expr

Implies Expr Expr

Equiv Expr Expr

Temporal (PathQuantified Expr)

data PathQuantified e

Path-quantified CTL expressions

Constructors

A (StateQuantified e)

E (StateQuantified e)

data StateQuantified e

State-quantified CTL expressions

Constructors

X e

G e

F e

U e e

parseExpr :: SourceName -> String -> Either ParseError Expr

Parse a CTL expressions from the given string.

The first parameter will be used to identify the source of the text in error messages.

This parser is compatible with the pretty printer of expressions, that is, pretty printed expressions will be parseable by this (unless they contain illegal identifiers for atomic propositions).

Model checking

check :: KripkeStructure String -> Expr -> Int -> Bool

Check if the given expression holds in the given state of the Kripke structure.

satisfyExpr :: KripkeStructure String -> Expr -> [State String]

Obtain all states that satisfy the given CTL expression.

satisfyExpr' :: KripkeStructure String -> Expr -> [Int]

Obtain the identifiers of all states that satisfy the given CTL expression.