Logic.Model

Description

Kripke structures

data KripkeStructure a

A Kripke structure is composed of a list of states and a list of transitions between such states. States are labeled with the atomic propositions that hold in it.

This particular kind of labeled transition system may be used as a model for temporal logics. In particular, it may be used for model checking.

This structure is polymorphic on the type of atomic propositions.

Constructors

KripkeStructure

Fields

states :: [State a]

List of labeled states of the Kripke structure

transitions :: [Transition a]

List of transitions of the Kripke structure

data State a

A state contains its unique identifier and a list of atomic propositions that hold in it.

Constructors

State Int [a]

data Transition a

A transition contains the identifiers of the source and target states.

Constructors

Transition

Fields

transitionId :: Int

source :: Int

target :: Int

transitionPayload :: [a]

Looking up states

stateIds :: KripkeStructure a -> [Int]

List of all state IDs from a given Kripke structure

lookupState :: Int -> KripkeStructure a -> Maybe (State a)

Finds the state with given ID in the given Kripke structure

getState :: Int -> KripkeStructure a -> State a

Gets the state with given ID in the given Kripke structure, fails if there is none.

Looking up transitions

transitionIds :: KripkeStructure a -> [Int]

List of all transition IDs on a given Kripke structure

lookupTransition :: Int -> KripkeStructure a -> Maybe (Transition a)

Finds the transition with given ID in the given Kripke structure.

getTransition :: Int -> KripkeStructure a -> Transition a

Gets the transition with given ID in the given Kripke structure, fails if there is none.

Lookup by adjacency

nextStates :: KripkeStructure a -> Int -> [Int]

Obtains the IDs of the states that are reachable by a single transition from the state with given ID.

precedes :: KripkeStructure a -> Int -> Int -> Bool

Tests if the second given state is reachable from the first by a single transition

prevStates :: KripkeStructure a -> Int -> [Int]

Obtains the IDs of the states from which the given state is reachable by a single transition.

follows :: KripkeStructure a -> Int -> Int -> Bool

Tests if the first given state is reachable from the second by a single transition

Utilities for labeled elements

class Element e where

Type class for elements that have a numeric identifier and a list of associated values.

Minimal complete definition

elementId, values

Associated Types

type Payload e :: *

Type of associated values.

Methods

elementId :: e -> Int

Obtain the numeric identifier of an element.

values :: e -> [Payload e]

Obtain the associated values of an element.

findById :: Element a => Int -> [a] -> Maybe a

Given a list of elements, find the element with the given identifier.