TypedGraph.DPO.GraphProcess

Documentation

data DoublyTypedGrammar a b

Constructors

DoublyTypedGrammar

Fields

singleTypedGrammar :: Grammar (TypedGraphMorphism a b)

The grammar typed over the double type graph

originalRulesWithMatches :: [NamedRuleWithMatches (TypedGraphMorphism a b)]

The rules of the original grammar, together with their matches in the double type graph

doubleType :: TypedGraph a b

The double type graph, typed over the simple type graph

originRelation :: Relation

The relation that shows the rule that originated each element (node or edge) in the grammar

concreteRelation :: Relation

The occurrence relation of the grammar (existencial relation + concrete conflicts and dependencies induced by NACs)

restrictRelation :: AbstractRelation

The set of restrictions given by the abstract conflicts and dependencies

occurrenceRelation :: [NamedProduction (TypedGraphMorphism a b)] -> Relation

generateDoublyTypedGrammar :: RuleSequence (TypedGraphMorphism a b) -> DoublyTypedGrammar a b

Given a rule sequence, it calculates its underlying doubly-typed graph grammar

uniqueOrigin :: [NamedProduction (TypedGraphMorphism a b)] -> Bool

findConcreteTrigger :: DoublyTypedGrammar a b -> Interaction -> (Interaction, RelationItem)

Given an doubly typed grammar and an interaction of the types ProduceForbid or DeleteForbid between two rules it returns the interaction together with the element that triggered the NAC involved in this conflict or dependency

calculateNacRelations :: DoublyTypedGrammar a b -> Set Interaction -> DoublyTypedGrammar a b

strictRelation :: [NamedProduction (TypedGraphMorphism a b)] -> Relation

creationAndDeletionRelation :: NamedProduction (TypedGraphMorphism a b) -> Relation

getElements :: DoublyTypedGrammar a b -> (Set RelationItem, Set RelationItem)

Given an doubly typed grammar, it returns a tuple (rs,es) where rs is the set of rule names in this grammar and es is the set of elements that appear in the rules

initialGraph :: DoublyTypedGrammar a b -> TypedGraph a b

Given an doubly typed grammar, it returns its initial graph

finalGraph :: DoublyTypedGrammar a b -> TypedGraph a b

Given an doubly typed grammar, it returns its final graph

emptyRestrictions :: DoublyTypedGrammar a b -> Bool

Checks whether the restrict relation of an doubly typed grammar is empty

Orphan instances

GenerateProcess (TypedGraphMorphism a b)

Methods

restrictMorphisms :: (TypedGraphMorphism a b, TypedGraphMorphism a b) -> (TypedGraphMorphism a b, TypedGraphMorphism a b)

restrictMorphism :: TypedGraphMorphism a b -> TypedGraphMorphism a b

typing :: (Derivation (TypedGraphMorphism a b), (TypedGraphMorphism a b, TypedGraphMorphism a b, TypedGraphMorphism a b)) -> Production (TypedGraphMorphism a b)

productionTyping :: (Production (TypedGraphMorphism a b), (TypedGraphMorphism a b, TypedGraphMorphism a b, TypedGraphMorphism a b)) -> Production (TypedGraphMorphism a b)

calculateProcess :: [Derivation (TypedGraphMorphism a b)] -> Process (TypedGraphMorphism a b)

calculateRulesColimit :: RuleSequence (TypedGraphMorphism a b) -> [NamedRuleWithMatches (TypedGraphMorphism a b)]

generateGraphProcess :: RuleSequence (TypedGraphMorphism a b) -> [(String, Production (TypedGraphMorphism a b))]