TypedGraph.DPO.OccurenceRelation

Documentation

data RelationItem

Constructors

Node NodeId

Edge EdgeId

Rule String

type Relation = Set (RelationItem, RelationItem)

type AbstractRelation = Set (AbstractType, (RelationItem, RelationItem), (RelationItem, RelationItem))

data AbstractType

Constructors

AbstractProduceForbid

AbstractDeleteForbid

isRuleAndElement :: (RelationItem, RelationItem) -> Bool

filterRulesOccurrenceRelation :: Relation -> Relation

filterElementsOccurrenceRelation :: Relation -> Relation

filterCreationRelation :: Relation -> Relation

filterDeletionRelation :: Relation -> Relation

isCreation :: (RelationItem, RelationItem) -> Bool

isDeletion :: (RelationItem, RelationItem) -> Bool

isNode :: RelationItem -> Bool

happensAfterAction :: Relation -> RelationItem -> String -> Bool

Tests wether an item appears after a rule in a given occurrence relation

happensBeforeAction :: Relation -> RelationItem -> String -> Bool

Tests wether an item appears before a rule in a given occurrence relation

relatedItens :: Relation -> (RelationItem, RelationItem) -> Bool

neverDeleted :: RelationItem -> Relation -> Bool

Given a relation item i and the deletion relation of an doubly typed grammar, it returns True if the item is deleted by some rule in this relation and False otherwise

present :: RelationItem -> Relation -> Bool

findOrder :: Relation -> Set RelationItem -> Maybe [RelationItem]

buildTransitivity :: Relation -> Relation

relationToString :: Relation -> String

restrictionToString :: AbstractRelation -> String