TypedGraph.Graph

Documentation

type TypedGraph a b = GraphMorphism (Maybe a) (Maybe b)

A typed graph is a morphism whose codomain is the type graph.

untypedGraph :: TypedGraph a b -> Graph (Maybe a) (Maybe b)

Obtain the untyped version of the typed graph

extractNodeType :: TypedGraph a b -> NodeId -> NodeId

extractEdgeType :: TypedGraph a b -> EdgeId -> EdgeId

typeGraph :: TypedGraph a b -> Graph (Maybe a) (Maybe b)

Obtain the type graph from a typed graph

null :: TypedGraph a b -> Bool

Test if the typed graph is empty

newTypedNodes :: TypedGraph a b -> [NodeId]

Infinite list of new node instances of a typed graph

newTypedEdges :: TypedGraph a b -> [EdgeId]

Infinite list of new edge instances of a typed graph

typedNodes :: TypedGraph a b -> [(NodeId, NodeId)]

Obtain a list of tuples (nodeId, typeId) for nodes in the graph.

typedEdges :: TypedGraph a b -> [(EdgeId, NodeId, NodeId, EdgeId)]

Obtain a list of tuples (edgeId, srcId, tgtId, typeId) for edges in the graph.

untypedNodes :: TypedGraph a b -> [NodeId]

Obtain the list of untyped nodes, i.e., the list of node ids from the typed graph domain

untypedEdges :: TypedGraph a b -> [EdgeId]

Obtain the list of untyped edges, i.e., the list of edge ids from the typed graph domain

Orphan instances

Cardinality (GraphMorphism a b)

Methods

cardinality :: GraphMorphism a b -> Int