TypedGraph.Morphism

Documentation

data TypedGraphMorphism a b

idMap :: GraphMorphism a b -> GraphMorphism a b -> TypedGraphMorphism a b

Creates a TypedGraphMorphism mapping the same elements of theirs codomains, from tgm1 to tgm2

isPartialInjective :: TypedGraphMorphism a b -> TypedGraphMorphism a b -> Bool

Test if a nac is partial injective (injective out of m)

invert :: TypedGraphMorphism a b -> TypedGraphMorphism a b

Invert a typed graph morphism

nodesFromDomain :: TypedGraphMorphism a b -> [NodeId]

Return the nodes in the domain of a given TypedGraphMorphism

edgesFromDomain :: TypedGraphMorphism a b -> [EdgeId]

Return the edges in the domain of a given TypedGraphMorphism

nodesFromCodomain :: TypedGraphMorphism a b -> [NodeId]

Return the nodes in the codomain of a given TypedGraphMorphism

edgesFromCodomain :: TypedGraphMorphism a b -> [EdgeId]

Return the edges in the codomain of a given TypedGraphMorphism

graphDomain :: TypedGraphMorphism a b -> Graph a b

Return the domain graph

graphCodomain :: TypedGraphMorphism a b -> Graph a b

Return the codomain graph

mapping :: TypedGraphMorphism a b -> GraphMorphism a b

applyNode :: TypedGraphMorphism a b -> NodeId -> Maybe NodeId

Given a TypedGraphMorphism t: G1 -> G2 and a node n in G1, it returns the node in G2 to which n gets mapped

applyEdge :: TypedGraphMorphism a b -> EdgeId -> Maybe EdgeId

Given a TypedGraphMorphism t: G1 -> G2 and an edge e in G1, it returns the edge in G2 to which e gets mapped

buildTypedGraphMorphism :: TypedGraph a b -> TypedGraph a b -> GraphMorphism a b -> TypedGraphMorphism a b

Given two TypedGraphs G1 and G2 and a simple GraphMorphism between them, it returns a TypedGraphMorphism from G1 to G2

checkDeletion :: Eq t => TypedGraphMorphism a b -> TypedGraphMorphism a b -> (TypedGraphMorphism a b -> t -> Maybe t) -> (TypedGraphMorphism a b -> [t]) -> t -> Bool

TODO: Find a better name for this function, that was repeated both here and in the GraphRule archive | Given the left-hand-side morphism of a rule l : K -> L, a match m : L -> G for this rule, an element e (that can be either a Node or an Edge) and two functions apply (for applying that element in a TypedGraphMorphism) and list (to get all the corresponding elements in the domain of m), it returns true if e is deleted by this rule for the given match

removeNodeFromDomain :: NodeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

Remove a node from the domain of a typed graph morphism

removeEdgeFromDomain :: EdgeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

Remove an edge from the domain of a typed graph morphism

removeNodeFromCodomain :: NodeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

Remove a node from the codomain of a typed graph morphism

removeEdgeFromCodomain :: EdgeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

Remove an edge from the domain of a typed graph morphism

applyNodeUnsafe :: TypedGraphMorphism a b -> NodeId -> NodeId

Given a TypedGraphMorphism tand a node n in the domain of t, return the node in the image of t to which n gets mapped or error in the case of undefined

applyEdgeUnsafe :: TypedGraphMorphism a b -> EdgeId -> EdgeId

Given a TypedGraphMorphism tand an edge e in the domain of t, return the edge in the image of t to which e gets mapped or error in the case of undefined

createEdgeOnDomain :: EdgeId -> NodeId -> NodeId -> EdgeId -> EdgeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

This function adds an edge e1 (with source s1, target t1 and type tp) to the domain of the typed graph morphism, and associate it to e2 It assumes s1, t1, e2, tp already exist, and that e1 does not exist.

createEdgeOnCodomain :: EdgeId -> NodeId -> NodeId -> EdgeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

This function adds an edge e2 (with source s2, target t2 and type tp) to the codomain of the typed graph morphism It assumes s2, t2, tp already exist, and that e2 does not exist.

createNodeOnDomain :: NodeId -> NodeId -> NodeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

This function adds a node n1 (type tp) to the domain of the typed graph morphism, and associate it to n2 It assumes n2 and tp already exist, and that n1 does not exist.

createNodeOnCodomain :: NodeId -> NodeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

This function adds a node n2 (type tp) to the codomain of the typed graph morphism It assumes tp already exist, and that n2 does not exist.

updateEdgeRelation :: EdgeId -> EdgeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

updates a typed graph morphism, mapping edge e1 to edge e2. It assumes both edges already exist.

updateNodeRelation :: NodeId -> NodeId -> NodeId -> TypedGraphMorphism a b -> TypedGraphMorphism a b

updates a typed graph morphism, mapping node n1 to node n2. It assumes both nodes already exist.

orphanTypedNodes :: TypedGraphMorphism a b -> [NodeId]

Given a TypedGraphMorphism, return its orphan nodes

orphanTypedEdges :: TypedGraphMorphism a b -> [EdgeId]

Given a TypedGraphMorphism, return its orphan edges

Orphan instances

FindMorphism (TypedGraphMorphism a b)

Methods

findMorphisms :: MorphismType -> Obj (TypedGraphMorphism a b) -> Obj (TypedGraphMorphism a b) -> [TypedGraphMorphism a b]

findMonomorphisms :: Obj (TypedGraphMorphism a b) -> Obj (TypedGraphMorphism a b) -> [TypedGraphMorphism a b]

findEpimorphisms :: Obj (TypedGraphMorphism a b) -> Obj (TypedGraphMorphism a b) -> [TypedGraphMorphism a b]

findIsomorphisms :: Obj (TypedGraphMorphism a b) -> Obj (TypedGraphMorphism a b) -> [TypedGraphMorphism a b]

findAllMorphisms :: Obj (TypedGraphMorphism a b) -> Obj (TypedGraphMorphism a b) -> [TypedGraphMorphism a b]

partialInjectiveMatches :: TypedGraphMorphism a b -> TypedGraphMorphism a b -> [TypedGraphMorphism a b]

EpiPairs (TypedGraphMorphism a b)

Methods

createJointlyEpimorphicPairs :: Bool -> Obj (TypedGraphMorphism a b) -> Obj (TypedGraphMorphism a b) -> [(TypedGraphMorphism a b, TypedGraphMorphism a b)]

createAllSubobjects :: Bool -> Obj (TypedGraphMorphism a b) -> [TypedGraphMorphism a b]

createJointlyEpimorphicPairsFromNAC :: DPOConfig -> Obj (TypedGraphMorphism a b) -> TypedGraphMorphism a b -> [(TypedGraphMorphism a b, TypedGraphMorphism a b)]

calculateCommutativeSquares :: Bool -> TypedGraphMorphism a b -> TypedGraphMorphism a b -> [(TypedGraphMorphism a b, TypedGraphMorphism a b)]

calculateCommutativeSquaresAlongMonomorphism :: (TypedGraphMorphism a b, Bool) -> (TypedGraphMorphism a b, Bool) -> [(TypedGraphMorphism a b, TypedGraphMorphism a b)]

AdhesiveHLR (TypedGraphMorphism a b)

Methods

calculatePushout :: TypedGraphMorphism a b -> TypedGraphMorphism a b -> (TypedGraphMorphism a b, TypedGraphMorphism a b)

hasPushoutComplement :: (MorphismType, TypedGraphMorphism a b) -> (MorphismType, TypedGraphMorphism a b) -> Bool

calculatePushoutComplement :: TypedGraphMorphism a b -> TypedGraphMorphism a b -> (TypedGraphMorphism a b, TypedGraphMorphism a b)

monomorphicPullback :: TypedGraphMorphism a b -> TypedGraphMorphism a b -> (TypedGraphMorphism a b, TypedGraphMorphism a b)