XML.GGXReader

Documentation

readGrammar :: String -> Bool -> MorphismsConfig -> IO (Grammar (TypedGraphMorphism a b), Grammar (RuleMorphism a b), [(String, Int)])

Reads the grammar in the XML, adds the needed minimal safety nacs to second order, and returns the grammar and a log

readGGName :: String -> IO String

readName :: String -> IO [String]

readNames :: String -> IO [(String, String)]

Reads the names of node/edge types and NACs, which are necessary when reexporting this grammar.

To lookup the name of a node type, use I ++ show nodeId as key, where nodeId is the ID of the node in the type graph. Lookup of edge types is analogous.

readTypeGraph :: String -> IO [ParsedTypeGraph]

readRules :: String -> IO [RuleWithNacs]

readGraphs :: String -> IO [(String, TypedGraph a b)]

readSequences :: Grammar (TypedGraphMorphism a b) -> String -> IO [(String, [GraphRule a b])]

readSequencesWithObjectFlow :: Grammar (TypedGraphMorphism a b) -> String -> IO [(String, [(String, GraphRule a b)], [ObjectFlow (TypedGraphMorphism a b)])]

instantiateRule :: TypeGraph a b -> RuleWithNacs -> GraphRule a b

instantiateSpan :: TypedGraph a b -> TypedGraph a b -> [Mapping] -> (TypedGraphMorphism a b, TypedGraphMorphism a b)

minimalSafetyNacsWithLog :: MorphismsConfig -> Grammar (RuleMorphism a b) -> (Grammar (RuleMorphism a b), [(String, Int)])

printMinimalSafetyNacsLog :: [(String, Int)] -> [String]