XML.GGXWriter

Documentation

type Grammars a b = (Grammar (TypedGraphMorphism a b), Grammar (RuleMorphism a b))

writeConfDepFile :: Bool -> MorphismsConfig -> Grammars a b -> String -> [(String, String)] -> String -> IO ()

Writes grammar, conflicts and dependencies (.cpx)

writeConflictsFile :: Bool -> MorphismsConfig -> Grammars a b -> String -> [(String, String)] -> String -> IO ()

Writes the grammar and the conflicts (.cpx)

writeDependenciesFile :: MorphismsConfig -> Grammars a b -> String -> [(String, String)] -> String -> IO ()

Writes the grammar and the dependencies (.cpx)

writeSndOderConfDepFile :: MorphismsConfig -> Grammars a b -> String -> [(String, String)] -> String -> IO ()

Writes grammar, second order conflicts and dependencies (.ggx)

writeSndOderConflictsFile :: MorphismsConfig -> Grammars a b -> String -> [(String, String)] -> String -> IO ()

Writes the grammar and the second order conflicts (.ggx)

writeSndOderDependenciesFile :: MorphismsConfig -> Grammars a b -> String -> [(String, String)] -> String -> IO ()

Writes the grammar and the second order dependencies (.ggx)

writeGrammarFile :: Grammars a b -> String -> [(String, String)] -> String -> IO ()

Writes only the grammar (.ggx)