Graph.Graph

Types

data Edge a

newtype EdgeId

Constructors

EdgeId Int

data Node a

newtype NodeId

Constructors

NodeId Int

data Graph a b

Constructors

Graph

Fields

nodeMap :: [(NodeId, Node a)]

edgeMap :: [(EdgeId, Edge b)]

Construction

empty :: Graph a b

Create an empty Graph.

build :: [Int] -> [(Int, Int, Int)] -> Graph a b

Build a Graph

Insertion

insertEdge :: EdgeId -> NodeId -> NodeId -> Graph a b -> Graph a b

Insert an edge e from src to tgt in graph g, without payload.

insertNode :: NodeId -> Graph a b -> Graph a b

Insert a node n in a graph g, without payload.

insertEdgeWithPayload :: EdgeId -> NodeId -> NodeId -> b -> Graph a b -> Graph a b

Insert an edge e from src to tgt in graph g with payload p.

insertNodeWithPayload :: NodeId -> a -> Graph a b -> Graph a b

Insert a node n in a graph g with payload p.

Delete/Update

removeEdge :: EdgeId -> Graph a b -> Graph a b

Remove edge e from g.

removeNode :: NodeId -> Graph a b -> Graph a b

If n exists in g, and there are no incident edges on it, remove it. Return g otherwise.

updateEdgePayload :: EdgeId -> Graph a b -> (b -> b) -> Graph a b

Update e's payload, applying f on it.

updateNodePayload :: NodeId -> Graph a b -> (a -> a) -> Graph a b

Update n's payload, applying f on it.

Extraction

edges :: Graph a b -> [EdgeId]

Return a list of all edge id's from from g.

nodes :: Graph a b -> [NodeId]

Return a list of all node id's from from g.

outgoingEdges :: Graph a b -> NodeId -> [EdgeId]

Return a list of all edges with n as a source node.

incomingEdges :: Graph a b -> NodeId -> [EdgeId]

Return a list of all edges with n as a target node.

edgesWithPayload :: Graph a b -> [(EdgeId, Maybe b)]

Return a list of all edge id's, together with their payloads.

incidentEdges :: Graph a b -> NodeId -> [EdgeId]

Return a list of all incident edges on n.

neighbourNodes :: Graph a b -> NodeId -> [NodeId]

Return a list of all neighbour nodes from n.

nodesOf :: Graph a b -> EdgeId -> Maybe (NodeId, NodeId)

Return a pair containing e's source and target nodes.

targetedNodes :: Graph a b -> NodeId -> [NodeId]

Return a list of all nodes that are target of any edge going out from n.

sourceNodes :: Graph a b -> NodeId -> [NodeId]

Return a list of all nodes that are source of any edge going into n.

nodesWithPayload :: Graph a b -> [(NodeId, Maybe a)]

Return a list of all node id's, together with their payloads.

Query

null :: Graph a b -> Bool

Test whether a graph is empty.

nodePayload :: Graph a b -> NodeId -> Maybe a

Return n's payload.

edgePayload :: Graph a b -> EdgeId -> Maybe b

Return e's payload.

sourceOf :: Graph a b -> EdgeId -> Maybe NodeId

Return e's source.

sourceOfUnsafe :: Graph a b -> EdgeId -> NodeId

Return e's source or error in the case of undefined

targetOf :: Graph a b -> EdgeId -> Maybe NodeId

Return e's target.

targetOfUnsafe :: Graph a b -> EdgeId -> NodeId

Return e's target or error in the case of undefined

extractSource :: Graph a b -> EdgeId -> NodeId

extractTarget :: Graph a b -> EdgeId -> NodeId

Predicates

isEdgeOf :: Graph a b -> EdgeId -> Bool

Test if e is an edge from graph g.

isNodeOf :: Graph a b -> NodeId -> Bool

Test if n is a node from graph g.

isAdjacentTo :: Graph a b -> NodeId -> NodeId -> Bool

Test if n1 and n2 are adjacent.

isIncidentTo :: Graph a b -> NodeId -> EdgeId -> Bool

Test if n is connected to edge e.

newNodes :: Graph a b -> [NodeId]

Infinite list of new node instances of a graph

newEdges :: Graph a b -> [EdgeId]

Infinite list of new edge instances of a graph