TypedGraph.Graph

Documentation

type TypedGraph a b = GraphMorphism a b

A typed graph is a morphism whose codomain is the type graph.

untypedGraph :: TypedGraph a b -> Graph a b

Obtain the untyped version of the typed graph

getEdgeType :: TypedGraph a b -> EdgeId -> EdgeId

Deprecated: Use extractEdgeType instead

getNodeType :: TypedGraph a b -> NodeId -> NodeId

Deprecated: Use extractNodeType instead

Given a TypedGraph and a Node in this graph, returns the type of the Node

extractNodeType :: TypedGraph a b -> NodeId -> NodeId

extractEdgeType :: TypedGraph a b -> EdgeId -> EdgeId

typeGraph :: TypedGraph a b -> Graph a b

Obtain the type graph from a typed graph

null :: TypedGraph a b -> Bool

Test if the typed graph is empty

newTypedNodes :: TypedGraph a b -> [NodeId]

Infinite list of new node instances of a typed graph

newTypedEdges :: TypedGraph a b -> [EdgeId]

Infinite list of new edge instances of a typed graph

nodesWithType :: TypedGraph a b -> [(NodeId, NodeId)]

Obtain a list of tuples (nodeId, typeId) for nodes in the graph.

edgesWithType :: TypedGraph a b -> [(EdgeId, NodeId, NodeId, EdgeId)]

Obtain a list of tuples (edgeId, srcId, tgtId, typeId) for edges in the graph.